Move static file stuff to it's own file
